fix calificacion por decmal

+++ Netmefy.Data/bt_ord_trabajo.cs
@@ -25,7 +25,7 @@ namespace Netmefy.Data
         public Nullable<int> tecnico_sk { get; set; }
         public System.DateTime fh_creacion { get; set; }
         public Nullable<System.DateTime> fh_cierre { get; set; }
-        public Nullable<int> calificacion { get; set; }
+        public Nullable<decimal> calificacion { get; set; }
         public int cliente_sk { get; set; }
         public int tipo { get; set; }
         public string descripcion { get; set; }
